Some changes as follows:\n 1. Secured the resources, i.e now authentication is required to perform any opeartion. Used spring security.\n 2. Some minor changes in usernotes.sql file. updated with insert scripts for sample data \n 3. End point URIs have been changed now and Readme document is updated. \n 4. Additional custom classes to leverage spring security

+++ src/main/java/com/opticores/resource/handler/NotesResourceHandler.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -13,21 +14,21 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.opticores.config.CustomUserdetails;
+import com.opticores.exception.AccessDeniedException;
+import com.opticores.exception.InvalidRequestException;
 import com.opticores.exception.NoEntityFoundException;
 import com.opticores.exception.UserNotFoundException;
 import com.opticores.model.ErrorMessage;
 import com.opticores.model.Note;
-import com.opticores.model.User;
 import com.opticores.service.NoteService;
 import com.opticores.service.UserService;
 
 /**
  * Main resource handler, handling clients requests for :
  * 
- * 1. fetch notes for a user 
- * 2. updating notes for a user 
- * 3. removing a user note 
- * 4. adding a new note
+ * 1. fetch notes for a user 2. updating notes for a user 3. removing a user
+ * note 4. adding a new note
  * 
  * 
  * @author anubhav
@@ -35,7 +36,7 @@ import com.opticores.service.UserService;
  */
 
 @RestController
-@RequestMapping(value = "/user")
+@RequestMapping(value = "/api/notes")
 public class NotesResourceHandler {
 
 	@Autowired
@@ -57,13 +58,10 @@ public class NotesResourceHandler {
 	 * 
 	 * @return a list of notes for a user
 	 */
-	@RequestMapping(value = "/{userid}/notes", produces = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.GET)
-	public ResponseEntity<List<Note>> getAllNotes(@PathVariable Integer userid) {
+	@RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.GET)
+	public ResponseEntity<List<Note>> getAllNotes() {
 
-		// 1. Check user existence
-		checkUserExistence(userid);
-
-		List<Note> notes = noteService.retrieveNotesForUser(userid);
+		List<Note> notes = noteService.retrieveNotesForUser(getUserId());
 
 		ResponseEntity<List<Note>> response = new ResponseEntity<List<Note>>(
 				notes, HttpStatus.OK);
@@ -78,7 +76,7 @@ public class NotesResourceHandler {
 	 * 
 	 * The 'endpoint' URI takes following form:
 	 * 
-	 * "/user/1/notes"
+	 * "/api/notes"
 	 * 
 	 * Request METHOD TYPE: POST
 	 * 
@@ -86,26 +84,22 @@ public class NotesResourceHandler {
 	 * @return a list of notes for a user
 	 * @throws UserNotFoundException
 	 */
-	@RequestMapping(value = "/{userid}/notes", consumes = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.POST)
-	public ResponseEntity<?> addNote(@RequestBody Note note,
-			@PathVariable Integer userid) {
-		
-		// 1. Check user existence
-		checkUserExistence(userid);
-		noteService.addNoteForUser(note, userid);
-		ResponseEntity<?> response= new ResponseEntity<>("Note created successfully",
-				HttpStatus.CREATED);
+	@RequestMapping(consumes = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.POST)
+	public ResponseEntity<?> addNote(@RequestBody Note note) {
+
+		noteService.addNoteForUser(note, getUserId());
+		ResponseEntity<?> response = new ResponseEntity<>(
+				"Note created successfully", HttpStatus.CREATED);
 
 		return response;
 
 	}
 
-	private void checkUserExistence(Integer userid)
-			throws UserNotFoundException {
-		User user = userService.getUserById(userid);
-		if (null == user) {
-			throw new UserNotFoundException();
-		}
+	private Integer getUserId() throws UserNotFoundException {
+
+		CustomUserdetails userDetails = (CustomUserdetails) SecurityContextHolder
+				.getContext().getAuthentication().getPrincipal();
+		return userDetails.getUserId();
 	}
 
 	/**
@@ -114,25 +108,26 @@ public class NotesResourceHandler {
 	 * 
 	 * The 'endpoint' URI takes following form:
 	 * 
-	 * "/user/1/notes"
+	 * "/api/notes"
 	 * 
 	 * Request METHOD TYPE: PUT
 	 * 
 	 * 
 	 * @return a list of notes for a user
 	 */
-	@RequestMapping(value = "/{userid}/notes", consumes = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.PUT)
-	public ResponseEntity<?> updateNote(@RequestBody Note note,
-			@PathVariable Integer userid) {
+	@RequestMapping(consumes = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.PUT)
+	public ResponseEntity<?> updateNote(@RequestBody Note note) {
 
-		// 1. Check user existence
-		checkUserExistence(userid);
+		Integer requestingUserId = getUserId();
+		checkUserAuthorization(note.getId(), requestingUserId);
 
-		noteService.updateNoteForUser(note, userid);
+		noteService.updateNoteForUser(note, requestingUserId);
 
 		return new ResponseEntity<>("Note updated successfully", HttpStatus.OK);
 
 	}
+
+	
 
 	/**
 	 * A function bound to an 'endpoint' to remove a resource( NOTES ) for a
@@ -140,20 +135,51 @@ public class NotesResourceHandler {
 	 * 
 	 * The 'endpoint' URI takes following form:
 	 * 
-	 * "/user/notes/2"
+	 * "/api/notes/2"
 	 * 
 	 * Request METHOD TYPE: DELETE
 	 * 
 	 * 
 	 * @return a list of notes for a user
 	 */
-	@RequestMapping(value = "/notes/{noteid}", method = RequestMethod.DELETE)
+	@RequestMapping(value = "/{noteid}", method = RequestMethod.DELETE)
 	public ResponseEntity<String> deleteNote(@PathVariable Integer noteid)
 			throws NoEntityFoundException {
-
+		
+		// 1. Retrieve the user id of the logged in user
+		Integer requestingUserId = getUserId();
+		
+		// 2. Check user access to perform this operation
+		checkUserAuthorization(noteid,requestingUserId);
+		
 		noteService.removeNoteForUser(noteid);
 
 		return new ResponseEntity<>("Note deleted successfully", HttpStatus.OK);
+
+	}
+	
+	
+	/**
+	 * This function basically checks if the requesting user is authorized to
+	 * perform an operation like UPDATE/DELETE a resource (NOTE)
+	 * 
+	 * if not authorized, it throws an appropriate runtime exception
+	 * 
+	 * @param note
+	 * @param requestingUserId
+	 * 
+	 */
+	private void checkUserAuthorization(Integer noteId, Integer requestingUserId) {
+
+		if (null == noteId) {
+			throw new InvalidRequestException();
+		}
+
+		Note originalNote = noteService.getNoteById(noteId);
+
+		if (!(originalNote.getUser().getId().equals(requestingUserId))) {
+			throw new AccessDeniedException();
+		}
 
 	}
